Move FullScreenAdapter to top level to fix regression
